Fix python bytecode generation
This was causing 'dnf' to segfault and 'mixer' to fail when this
package happens to be updated in a way that the '.so' wont be
"compatible" with the previous version. Python won't notice the change
and will try to use the '.pyc' file it created based on the previous
version.

After digging logs and a lot of discussions, it was a consensus that,
nowadays, it is no longer necessary to disable the post_install hook
because debug info is properly handled now.

This patch restore the default hook and removes the additions to
'%install' that were made to keep the desired job of the default hook in
place.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>