Backport bugfix for rpmbuild
There were multiple cases of the team trying to build 'mesa' with LTO
enabled, but Koji failed the build task because a corrupt debuginfo RPM
was generated.

This issue was addressed in the 4.13.0.1 release, so backport that fix.
When we eventually move to a newer rpm release, we can drop this patch.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>